fix(benchmarks): break synthesizer circular dependency
Extract candidate/review types and parsers into generationCandidate so
dataQualityVerifier no longer runtime-imports datasetGenerator.
