+++ ts/packages/benchmarks/src/translationBench/synthesizer/index.ts
@@ -8,6 +8,7 @@ export * from "./generationSupport.js";
 export * from "./benchmark.js";
 export * from "./sourceAdapter.js";
 export * from "./sourceBuilder.js";
+export * from "./generationCandidate.js";
 export * from "./datasetGenerator.js";
 export * from "./dataQualityVerifier.js";
 export * from "./synthesizerPrompts.js";
